Add SetHeaders and update interface to include this new method.
The purpose of the setter is to allow the setting of the headers at any time
during the life of the api service. The motivation was to allow a service
created with a given session, the chance to modify the headers, since the caller
would not have been able to use NewWithHeaders.

Even though it would be convenient to simply export Headers in the aide struct
definition to avoid modifying the interface, and thus causing breaking changes,
this won't do. It must be added to the interface if anybody wants to be able to
both use the interface... and set the headers.

Any packages implementing apigatewayiface.Service will need to add the new
method. Since this is almost certainly used as a test interface, you can
probably just regenerate your mocks.